esercizi completati.

+++ admin/admin.js
@@ -16,26 +16,35 @@ let changeBrand = document.querySelector("#change-brand")
 // Dove appendere il DOM
 let productsContainer = document.querySelector('#products-container > div')
 
-
+let formChangeBtn = document.querySelector('#form-change')
 
 // Tasto conferma cancella prodotto
 let confirmDeleteBtn = document.querySelector('#confirm-delete-btn')
 
+// tasto chiusura offcanvas
+let closeBtn = document.querySelectorAll('.closeX')
+
+
+
+let arrayProducts = []
 
 
 window.onload = async function () {
 
     productsContainer.innerHTML = /*html*/ `
     <div id="loader" class="loading p-5 mt-5">
-        <div></div>
-        <div></div>
-        <div></div>
+    <div></div>
+    <div></div>
+    <div></div>
     </div>
-   
-`
+    
+    `
     try {
-        allProducts = await getProducts()
+        let allProducts = await getProducts()
 
+            .finally(() => {
+                productsContainer.querySelector("#loader").remove();
+            })
         console.log(allProducts)
 
 
@@ -44,19 +53,17 @@ window.onload = async function () {
         productsContainer = document.querySelector('#products-container > div')
 
 
+
+
+        arrayProducts = allProducts
         // Mostro tutti i prodotti nel DOM
-        displayProducts(allProducts)
-        .finally(() => {
-            productsContainer.querySelector("#loader").remove();
-        })
+        displayProducts(arrayProducts)
 
     } catch (error) {
         console.log(error)
     }
 
-
 }
-
 
 
 
@@ -71,7 +78,7 @@ async function getProducts() {
         })
         const jsonData = await response.json()
 
-      
+
 
         return jsonData
     } catch (error) {
@@ -86,27 +93,27 @@ async function getProducts() {
 function displayProducts(data) {
 
     productsContainer.innerHTML = data.map(({ _id, name, price, imageUrl, description, brand }) => /*html*/`
-                                    <div id="_${_id}" class="row p-2 bg-white border rounded">
-                                        <div class="col-md-3 mt-1">
-                                            <img class="img-fluid img-responsive rounded product-image"src="${imageUrl}">
-                                        </div>
-                                        <div class="col-md-6 mt-1">
-                                            <h5>${name}</h5>
-                                            <p>${brand}</p>
-                                            <p class="text-justify text-truncate para mb-0">${description}<br><br></p>
-                                        </div>
-                                        <div class="align-items-center align-content-center col-md-3 border-left mt-1">
-                                            <div class="d-flex flex-row align-items-center">
-                                             <h4 class="mr-1">${price}€</h4>
-                                            </div>
-                                            <div class="d-flex flex-column mt-4">
-                                                <button class="btn btn-success btn-sm" data-bs-toggle="offcanvas"
-                                                data-bs-target="#change-product" aria-controls="change-product" onclick="fillForm(${price})">Change</button>
-                                                <button class="btn btn-danger btn-sm mt-2" type="button" onclick="deleteProduct('${_id}')">Delete</button>
-                                            </div>
-                                        </div>
-                                    </div>
-                    `
+        <div id="_${_id}" class="row p-2 bg-white border rounded my-3">
+        <div class="col-md-3 mt-1">
+        <img class="img-fluid img-responsive rounded product-image"src="${imageUrl}">
+        </div>
+        <div class="col-md-6 mt-1">
+        <h5>${name}</h5>
+        <p>${brand}</p>
+        <p class="text-justify para mb-0">${description}<br><br></p>
+        </div>
+        <div class="align-items-center align-content-center col-md-3 border-left mt-1">
+        <div class="d-flex flex-row align-items-center">
+        <h4 class="mr-1">${price}€</h4>
+        </div>
+        <div class="d-flex flex-column mt-4">
+        <button class="btn btn-success btn-sm" data-bs-toggle="offcanvas"
+        data-bs-target="#change-product" aria-controls="change-product" onclick="fillForm('${_id}')">Edit</button>
+        <button class="btn btn-danger btn-sm mt-2" type="button" onclick="deleteProduct('${_id}')">Delete</button>
+        </div>
+        </div>
+        </div>
+        `
     ).join('')
 
 
@@ -132,7 +139,12 @@ async function addProduct(event) {
             brand: brand.value
         })
     })
+        .finally(() => {
+            for (const button of closeBtn) {
+                button.click()
 
+            }
+        })
     title = document.querySelector("#titleInput")
     price = document.querySelector("#priceInput")
     imageUrl = document.querySelector("#imageUrl")
@@ -155,30 +167,40 @@ async function addProduct(event) {
 }
 
 // funzione che apre modale per modificare prodotti
-function fillForm(id, name, price, imageUrl, description, brand) {
-    // idItem.value = id
-    changeTitle.value = name
-    changePrice.value = price
-    changeImageUrl.value = imageUrl
-    changeDescription.value = description
-    changeBrand.value = brand
+function fillForm(id) {
+    console.log(arrayProducts)
+    const itemSelected = arrayProducts.find(item => item._id === id)
+    console.log(itemSelected)
+
+    idItem.innerHTML = itemSelected._id
+    changeTitle.value = itemSelected.name
+    changePrice.value = itemSelected.price
+    changeImageUrl.value = itemSelected.imageUrl
+    changeDescription.value = itemSelected.description
+    changeBrand.value = itemSelected.brand
+
+
+
 
 
 }
 
+
 // Funzione che modifica prodotto
-async function changeProduct(event, id) {
+async function changeProduct(event) {
+
+    formChangeBtn.classList.add('pe.none')
     //  aggiunge pe none per non far cliccare
     event.preventDefault()
 
-    const response = await fetch('https://striveschool-api.herokuapp.com/api/product/' + id, {
+    const response = await fetch('https://striveschool-api.herokuapp.com/api/product/' + idItem.innerHTML, {
         method: "PUT",
         headers: {
             "Content-type": "application/json",
             "Authorization": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI2NTFhZTkzYjk2OTNhMDAwMThiOWRlYjkiLCJpYXQiOjE2OTYyNjI0NTksImV4cCI6MTY5NzQ3MjA1OX0.zDkQ9wBotUv_FGzLQHZSRMthIEqAPb40vnYwaa5RBL4"
         },
         body: JSON.stringify({
-            _id: idItem.value,
+            _id: idItem.innerHTML,
             name: changeTitle.value,
             price: changePrice.value,
             imageUrl: changeImageUrl.value,
@@ -186,7 +208,12 @@ async function changeProduct(event, id) {
             brand: changeBrand.value
         })
     })
-    console.log(idItem.value)
+        .finally(() => {
+            for (const button of closeBtn) {
+                button.click()
+
+            }
+        })
 
     idItem = document.querySelector('#id-item')
     changeTitle = document.querySelector('#change-title')
@@ -200,13 +227,15 @@ async function changeProduct(event, id) {
 
         const data = await getProducts()
         displayProducts(data)
+        confirm('Successfully added!')
 
         for (const field of [idItem, changeTitle, changePrice, changeImageUrl, changeDescription]) {
             field.value = ''
         }
     } else {
-        console.error("Cannot send")
+        console.error("Something got wrong")
     }
+
 }
 
 
@@ -217,6 +246,9 @@ async function changeProduct(event, id) {
 // Funzione che elimina prodotto
 async function deleteProduct(id) {
 
+    if (!confirm("Are you sure to delete?")) {
+        return
+    }
     const response = await fetch('https://striveschool-api.herokuapp.com/api/product/' + id, {
         method: "DELETE",
         headers: {
@@ -225,9 +257,6 @@ async function deleteProduct(id) {
         }
 
     })
-    if (!confirm("Are you sure to delete?")) {
-        return
-    }
     if (response.ok) {
 
         const data = await getProducts()
@@ -238,6 +267,7 @@ async function deleteProduct(id) {
     }
 
 }
+
